Add working asteroids and reduce target FPS to 30.

+++ src/utils/vectors.rs
@@ -1,5 +1,5 @@
 pub struct Vector 
 {
-    pub x: i16,
-    pub y: i16,
+    pub x: i32,
+    pub y: i32,
 }
